chore: code cleanup and refactoring
Move legal move generation code to new file
