+++ board/lib.rs
@@ -19,6 +19,7 @@ pub mod board_state;
 pub mod definitions;
 pub mod fen;
 pub mod file;
+pub mod legal_move_generation;
 pub mod magics;
 pub mod move_generation;
 pub mod move_history;
